no crashing when typing strings into int values
in the gambling menu, you may quit by the muscle memory method, this doesn't just close the game, it doesn't save your progress either. This can be very annoying if you played for a long session. To fix this, i made it that the program does nothing if there is an error instead of making an error. You are welcome!